easywsclient for Windows
Revised a source file to be able to compile it in Linux and Windows.

Revised a description error of the include sentence for Windows.

followed the modification of the original version.

preparing for pull request.

preparing for pull request.

Convert tabs to spaces.
Add delete ws; to 'example-client.cpp'.

try2

try3